Updated: - generation.manager: commented unneeded function 'find_associated_bus_ehighway' - geographics.manager: improved a bit 'match_points_to_region' - load.manager: added a function 'retrieve_load_data_per_country' to get load directly from countries list - network_builder.nuclear: removed dependency on e-highway - network_builder.res:  added limit_max_cap argument to add_generators - postprocessing.resite_plotly and postprocessing.resite_results - resite.models.pyomo: changed hourly formulation
Divided:
- topologies.ehighway.py into topologies.ehighways.py (specific function for ehighway topology) and topologies.manager.py (generic functions regarding topology)

Created:
- topologies.tyndp2018.py: functions to deal with the new country-based topology
- sizing.tyndp2018.main.py: script for building tyndp2018 network (not totally working yet)

+++ src/resite/resite.py
@@ -72,7 +72,7 @@ class Resite:
 
     def __del__(self):
         """If self.keep_files is false, remove all outputs created during the run."""
-        if not self.keep_files:
+        if not self.keep_files and hasattr(self, "output_folder"):
             rmtree(self.output_folder)
 
     # TODO: this is pretty messy - find a way to clean it up
